+++ CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.19.0] - 2025-12-19
 
 - Display indexation status in indexes table
+- The facet limit and display limit can be set individually
+- Add a setting to add checkboxes to facets in order to combine them and reload the filtered search results
 
 ## [0.18.2] - 2025-11-26